pruned tests, files
